+++ client/src/views/guild/OwnGuild.vue
